fix(runtime): resolve documentation-review findings VDOC-129..225
Harden the runtime and its machine-readable surface across the
documentation-review findings batch: trap-discipline guards and handle
validators (Random, Perlin, BinaryBuffer, Mat4), PTY/process PATH-search
unification with exec-status surfacing, Windows wide-API/version/env
corrections, terminal parameter clamps, a shared monotonic-clock fallback
ratchet so the POSIX CLOCK_REALTIME path stays non-decreasing, an
acquire/release fix for the Windows QPC-frequency data race, and a new
Option-returning DateTime.TryFromParts that disambiguates the -1 sentinel
from the valid pre-epoch instant.

Correct dump-runtime-api fallibility, ownership, and nullability metadata
for the IO, Math, and System surfaces, pin the corrected contracts in the
agent_cli test, and regenerate the runtime reference docs.

tests: added/extended runtime regression suites (handle validation,
watcher, ide workspace, savedata, spline, perlin, bigint, quat, shutdown,
args, exec, terminal color, clock ratchet, stopwatch concurrency, datetime
option) and agent_cli contract pins; full default suite green.

+++ src/runtime/core/rt_string.h
@@ -255,7 +255,7 @@ rt_string rt_str_d_alloc(double v);
 /// @brief Allocate a runtime string from a float value using standard formatting.
 /// @param v Single-precision value to convert.
 /// @return Newly allocated runtime string.
-rt_string rt_str_f_alloc(float v);
+rt_string rt_str_f_alloc(double v);
 
 /// @brief Allocate a runtime string from a 32-bit integer using standard formatting.
 /// @param v 32-bit integer value to convert.
